Restructure รอบที่ 2
