+++ app/blueprints/api/__init__.py
@@ -1,4 +1,4 @@
-# app/blueprints/auth/__init__.py
+# app/blueprints/api/__init__.py
 from flask import Blueprint
-bp = Blueprint("auth", __name__, url_prefix="/auth") # <-- เพิ่ม url_prefix ที่นี่
-from . import routes  # <-- เพิ่มบรรทัดนี้
+bp = Blueprint("api", __name__, url_prefix="/api")
+from . import routes
