+++ app/blueprints/owner/route_properties.py
@@ -8,7 +8,7 @@ from app.forms.owner import PropertyForm
 from app.forms.upload import UploadImageForm, ReorderImagesForm, EmptyForm
 from app.models.property import Property, PropertyImage, Amenity
 from app.models.approval import ApprovalRequest
-from app.extensions import db
+from app.core.extensions import db
 
 # Policy import
 try:
