+++ app/models/review.py
@@ -1,5 +1,5 @@
 # app/models/review.py
-from app.extensions import db
+from app.core.extensions import db
 from datetime import datetime
 
 class Review(db.Model):
